+++ pkg/template/scripts/prometheus.go
@@ -26,7 +26,7 @@ import (
 // PrometheusScript represent the data to generate Prometheus config
 type PrometheusScript struct {
 	IP        string
-	Port      uint64
+	Port      int
 	DeployDir string
 	DataDir   string
 	LogDir    string
@@ -45,7 +45,7 @@ func NewPrometheusScript(ip, deployDir, dataDir, logDir string) *PrometheusScrip
 }
 
 // WithPort set Port field of PrometheusScript
-func (c *PrometheusScript) WithPort(port uint64) *PrometheusScript {
+func (c *PrometheusScript) WithPort(port int) *PrometheusScript {
 	c.Port = port
 	return c
 }
