+++ pkg/template/scripts/tidb.go
@@ -26,8 +26,8 @@ import (
 // TiDBScript represent the data to generate TiDB config
 type TiDBScript struct {
 	IP         string
-	Port       uint64
-	StatusPort uint64
+	Port       int
+	StatusPort int
 	DeployDir  string
 	LogDir     string
 	NumaNode   string
@@ -46,13 +46,13 @@ func NewTiDBScript(ip, deployDir, logDir string) *TiDBScript {
 }
 
 // WithPort set Port field of TiDBScript
-func (c *TiDBScript) WithPort(port uint64) *TiDBScript {
+func (c *TiDBScript) WithPort(port int) *TiDBScript {
 	c.Port = port
 	return c
 }
 
 // WithStatusPort set StatusPort field of TiDBScript
-func (c *TiDBScript) WithStatusPort(port uint64) *TiDBScript {
+func (c *TiDBScript) WithStatusPort(port int) *TiDBScript {
 	c.StatusPort = port
 	return c
 }
